feat(core/sdbackup): remove sd_backup from apps
- the sd_backup won't be triggered as a standalone app by protobuf so
  there is no reason to have the file there. Rather move the function
  closer to their usage.
